adds a simple script to calculate using python
adds a script to lets user input two numbers and a operator to do simple math.
The code prone to errors if input is not two numbers and a operator.
The program also loops back to beginning so you can caluculate as many times you want,
and if you want to exit you just force an error.